refactor: test runner logger and reporters modules
