+++ packages/test-runner-logger/README.md
@@ -1,0 +1,5 @@
+# Test Runner Logger
+
+Default logger for web test runner
+
+See [our website](https://modern-web.dev/docs/test-runner/overview/) for full documentation.
