+++ packages/test-runner-logger/package.json
@@ -1,0 +1,52 @@
+{
+    "name": "@web/test-runner-logger",
+    "version": "0.1.0",
+    "publishConfig": {
+      "access": "public"
+    },
+    "description": "Default logger for @web/test-runner",
+    "license": "MIT",
+    "repository": {
+      "type": "git",
+      "url": "https://github.com/modernweb-dev/web.git",
+      "directory": "packages/test-runner-logger"
+    },
+    "author": "modern-web",
+    "homepage": "https://github.com/modernweb-dev/web/tree/master/packages/test-runner-logger",
+    "main": "dist/index.js",
+    "exports": {
+      ".": {
+        "import": "./index.mjs",
+        "require": "./dist/index.js"
+      }
+    },
+    "engines": {
+      "node": ">=12.0.0"
+    },
+    "scripts": {
+      "build": "tsc",
+      "test": "mocha test/**/*.test.ts --require ts-node/register --reporter dot",
+      "test:watch": "mocha test/**/*.test.ts --require ts-node/register --watch --watch-files src,test --reporter dot"
+    },
+    "files": [
+      "*.d.ts",
+      "*.js",
+      "*.mjs",
+      "dist",
+      "src"
+    ],
+    "keywords": [
+      "web",
+      "test",
+      "runner",
+      "testrunner",
+      "logger"
+    ],
+    "dependencies": {
+        "@web/test-runner-core": "^0.10.0"
+    },
+    "devDependencies": {
+      "@web/test-runner-playwright": "^0.8.7"
+    }
+  }
+
