+++ packages/test-runner-reporters/README.md
@@ -1,0 +1,5 @@
+# Test Runner Reporters
+
+Utils and default reporters for web test runner
+
+See [our website](https://modern-web.dev/docs/test-runner/reporters/overview/) for full documentation.
