+++ packages/test-runner-reporters/src/dotReporter.ts
@@ -1,53 +1,39 @@
 import type { Reporter, ReporterArgs, TestSuiteResult } from '@web/test-runner-core';
+import { TestRunnerLogger } from '@web/test-runner-logger';
 
 import { reportTestsErrors } from './reportTestsErrors.js';
 import { reportTestFileErrors } from './reportTestFileErrors.js';
-
-import { TestRunnerLogger } from '../logger/TestRunnerLogger.js';
-
-interface Options {
-  flatten?: boolean;
-}
 
 const color =
   ([x, y]: [number, number]) =>
   (z: string) =>
     `\x1b[${x}m${z}\x1b[${y}m${reset}`;
 const reset = `\x1b[0m\x1b[0m`;
-const green = color([32, 89]);
 const red = color([31, 89]);
 
-/** Test reporter that summarizes all test for a given run */
-export function summaryReporter(opts: Options): Reporter {
-  const { flatten = false } = opts ?? {};
+/** Test reporter that summarizes all test for a given run using dot notation */
+export function dotReporter(): Reporter {
   let args: ReporterArgs;
   let favoriteBrowser: string;
 
   const logger = new TestRunnerLogger();
 
-  function log(name: string, passed: boolean, prefix = '  ') {
-    const sign = passed ? green('✓') : red('𐄂');
-    logger.log(flatten ? `${sign} ${prefix} ${name}` : `${prefix}  ${sign} ${name}`);
+  function log(passed: boolean) {
+    logger.log(passed ? '.' : red('x'));
   }
 
-  function logResults(results?: TestSuiteResult, prefix?: string) {
+  function logResults(results?: TestSuiteResult) {
     for (const result of results?.tests ?? []) {
-      log(result.name, result.passed, prefix);
+      log(result.passed);
     }
 
     for (const suite of results?.suites ?? []) {
-      logSuite(suite, prefix);
+      logSuite(suite);
     }
   }
 
-  function logSuite(suite: TestSuiteResult, parent?: string) {
-    const pref = flatten ? `${parent ? `${parent}` : ''} ${suite.name}` : `${parent ?? ''} `;
-
-    if (!flatten) {
-      logger.log(`${pref}${suite.name}`);
-    }
-
-    logResults(suite, pref);
+  function logSuite(suite: TestSuiteResult) {
+    logResults(suite);
   }
 
   return {
@@ -63,7 +49,6 @@ export function summaryReporter(opts: Options): Reporter {
     reportTestFileResults({ sessionsForTestFile: sessions }) {
       for (const session of sessions) {
         logResults(session.testResults);
-        logger.log('');
       }
     },
