+++ packages/test-runner-reporters/src/getTestProgress.ts
