+++ packages/test-runner-reporters/src/renderProgressBar.ts
