+++ packages/test-runner-reporters/src/utils/getFailedOnBrowsers.ts
