+++ packages/test-runner-reporters/src/utils/getPassedFailedSkippedCount.ts
