+++ packages/test-runner-reporters/tsconfig.json
@@ -29,43 +29,10 @@
       "path": "../test-runner-mocha/tsconfig.json"
     },
     {
-      "path": "../test-runner-chrome/tsconfig.json"
-    },
-    {
-      "path": "../config-loader/tsconfig.json"
-    },
-    {
-      "path": "../dev-server-rollup/tsconfig.json"
-    },
-    {
       "path": "../test-runner-playwright/tsconfig.json"
     },
     {
-      "path": "../test-runner-webdriver/tsconfig.json"
-    },
-    {
       "path": "../test-runner-logger/tsconfig.json"
-    },
-    {
-      "path": "../polyfills-loader/tsconfig.json"
-    },
-    {
-      "path": "../dev-server-esbuild/tsconfig.json"
-    },
-    {
-      "path": "../dev-server-legacy/tsconfig.json"
-    },
-    {
-      "path": "../dev-server/tsconfig.json"
-    },
-    {
-      "path": "../test-runner-commands/tsconfig.json"
-    },
-    {
-      "path": "../test-runner-reporters/tsconfig.json"
-    },
-    {
-      "path": "../test-runner-saucelabs/tsconfig.json"
     }
   ],
   "include": [
