+++ packages/test-runner/package.json
@@ -79,7 +79,9 @@
     "@web/test-runner-chrome": "^0.10.7",
     "@web/test-runner-commands": "^0.6.0",
     "@web/test-runner-core": "^0.10.26",
+    "@web/test-runner-logger": "^0.1.0",
     "@web/test-runner-mocha": "^0.7.5",
+    "@web/test-runner-reporters": "^0.1.0",
     "camelcase": "^6.2.0",
     "command-line-args": "^5.1.1",
     "command-line-usage": "^6.1.1",
