+++ packages/test-runner/src/index.ts
@@ -2,9 +2,7 @@ import type { TestRunnerConfig as FullTestRunnerConfig } from './config/TestRunn
 
 export * from '@web/test-runner-core';
 export { chromeLauncher } from '@web/test-runner-chrome';
+export { defaultReporter, summaryReporter, dotReporter } from '@web/test-runner-reporters';
 
 export { startTestRunner } from './startTestRunner';
-export { defaultReporter } from './reporter/defaultReporter';
-export { summaryReporter } from './reporter/summaryReporter';
-export { dotReporter } from './reporter/dotReporter';
 export type TestRunnerConfig = Partial<FullTestRunnerConfig>;
