+++ tsconfig.json
@@ -41,10 +41,13 @@
       "path": "./packages/test-runner-playwright/tsconfig.json"
     },
     {
-      "path": "./packages/dev-server-esbuild/tsconfig.json"
+      "path": "./packages/test-runner-webdriver/tsconfig.json"
     },
     {
-      "path": "./packages/test-runner-webdriver/tsconfig.json"
+      "path": "./packages/test-runner-logger/tsconfig.json"
+    },
+    {
+      "path": "./packages/dev-server-esbuild/tsconfig.json"
     },
     {
       "path": "./packages/dev-server-legacy/tsconfig.json"
@@ -56,6 +59,9 @@
       "path": "./packages/test-runner-commands/tsconfig.json"
     },
     {
+      "path": "./packages/test-runner-reporters/tsconfig.json"
+    },
+    {
       "path": "./packages/test-runner-saucelabs/tsconfig.json"
     },
     {
@@ -63,9 +69,6 @@
     },
     {
       "path": "./packages/rollup-plugin-polyfills-loader/tsconfig.json"
-    },
-    {
-      "path": "./packages/test-runner-selenium/tsconfig.json"
     },
     {
       "path": "./packages/rollup-plugin-copy/tsconfig.json"
@@ -87,6 +90,9 @@
     },
     {
       "path": "./packages/test-runner-puppeteer/tsconfig.json"
+    },
+    {
+      "path": "./packages/test-runner-selenium/tsconfig.json"
     },
     {
       "path": "./packages/test-runner-browserstack/tsconfig.json"
