+++ workspace-packages.mjs
@@ -18,6 +18,8 @@ const packages = [
   { name: 'dev-server-storybook', type: 'ts', environment: 'node' },
   { name: 'test-runner', type: 'ts', environment: 'node' },
   { name: 'test-runner-core', type: 'ts', environment: 'node' },
+  { name: 'test-runner-logger', type: 'ts', environment: 'node' },
+  { name: 'test-runner-reporters', type: 'ts', environment: 'node' },
   { name: 'test-runner-chrome', type: 'ts', environment: 'node' },
   { name: 'test-runner-puppeteer', type: 'ts', environment: 'node' },
   { name: 'test-runner-playwright', type: 'ts', environment: 'node' },
